+++ config/settings.py
@@ -297,6 +297,7 @@ SPECTACULAR_SETTINGS = {
         "FlawSource": "osidb.models.flaw.source.FlawSource",
         "FlawReferenceType": "osidb.models.flaw.reference.FlawReference.FlawReferenceType",
         "TrackerType": "osidb.models.Tracker.TrackerType",
+        "FlawLabelType": "osidb.models.flaw.label.FlawLabel.FlawLabelType",
     },
     "ENUM_GENERATE_CHOICE_DESCRIPTION": False,
     "COMPONENT_SPLIT_REQUEST": True,